keep backwards compat junk in mycroft module
